@@ -1,6 +1,7 @@
-'use strict';
-const $ = require('jquery');
+// interesting jquery import thanks to babel
+import $ from 'jquery';
+// styles
 import './footer.css';
 
-$('body').append('<footer></footer>');
-$('footer').append('<p>Copyright - Holberton School</p>');
+// append elements
+$('body').append(`<p>Copyright - Holberton School</p>`);
