@@ -1,21 +1,24 @@
-'use strict';
+// interesting jquery import thanks to babel
+import $ from 'jquery';
+// import lodash
+import _ from 'lodash';
+// styles
 import '../css/main.css';
-const $ = require('jquery');
-const _ = require('lodash');
 
-$('body').append('<p>Holberton Dashboard</p>');
-$('body').append('<p>Dashboard data for the students</p>');
-$('body').append('<button>Click here to get started</button>');
-$('body').append('<p id="count"></p>');
-$('body').append('<p>Copyright - Holberton School</p>');
+// append elements
+$('body').append(`<div id="logo"></div>`);
+$('body').append(`<p>Holberton Dashboard</p>`);
+$('body').append(`<p>Dashboard data for the students</p>`);
+$('body').append(`<button>Click here to get started</button>`);
+$('body').append(`<p id="count"></p>`);
+$('body').append(`<p>Copyright - Holberton School</p>`);
 
-const updateCounter = () => {
-  let times = $('#count').html() || 0;
-  $('button').on('click', () => {
-    times++;
-    $('#count').html(`${times} clicks on the button`);
-  });
-};
+// bind lodash's debounce to the button
+$('button').on('click', _.debounce(updateCounter, 500));
 
-_.debounce(updateCounter, 500);
-updateCounter();
+let count  = 0;
+// function to update the counter
+function updateCounter() {
+	count++;
+	$('#count').text(`${count} clicks on the button`);
+}
